% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_results}
\alias{classification_results}
\title{Classification results}
\source{
See the \code{example-data} directory in the package with code that is a variation
of the analysis shown at \url{https://www.tidymodels.org/start/case-study/}.
}
\value{
\item{classification_results}{a tibble}
}
\description{
These data are a variation of a case study at \code{tidymodels.org} where a
penalized regression model was used for a binary classification task. The
outcome metrics in \code{classification_results} are the areas under the ROC and
PR curve, log-likelihood, and the number of predictors selected for a given
amount of penalization. Two tuning parameters, \code{mixture} and \code{penalty}, were
varied across 300 conditions.
}
\examples{

data(classification_results)
}
\keyword{datasets}
