% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desirability2-package.R
\docType{package}
\name{desirability2-package}
\alias{desirability2}
\alias{desirability2-package}
\title{In-line desirability functions}
\description{
Desirability functions (Derringer and Suich, 1980,
\url{doi:10.1080/00224065.1980.11980968}) can be used for multivariate
optimization.
}
\details{
See:
\itemize{
\item \code{\link[=d_max]{d_max()}} for individual desirability functions.
\item \code{\link[=d_overall]{d_overall()}} for combining functions into an overall objective function.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://desirability2.tidymodels.org}
  \item \url{https://github.com/tidymodels/desirability2}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
