% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imbalance_example}
\alias{imbalance_example}
\title{Imbalanced classification results}
\value{
\item{imbalance_example}{an object with class \code{tune_results}}
}
\description{
This is an object produced by \code{\link[tune:tune_grid]{tune::tune_grid()}} for a classification data
set with a moderate class imbalance (18\%) in the first class. The data were
simulated. The code to produce the data and analysis is in the source
package’s \code{inst} directory.
}
\details{
The model was a single-layer neural network with six tuning parameters
chosen for optimization. One parameter, \verb{class_weights,} controls how much
influence the minority class has on the objective function during training.
This can help trade off the true positive and true negative rates to our
needs.

The goal would be to have good calibration (as measured by \code{brier_class})
and improved \code{sensitivity}, but not without a significant loss of
\code{specificity}.
}
\keyword{datasets}
