% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{make_desirability_cols}
\alias{make_desirability_cols}
\title{Make desirability columns from a desirability object}
\usage{
make_desirability_cols(x, dat, call = rlang::caller_env())
}
\arguments{
\item{x}{An object produced by \code{\link[=desirability]{desirability()}}}

\item{dat}{A data frame}

\item{call}{The execution environment of a currently running function.}
}
\value{
An updated version of \code{dat}
}
\description{
Make desirability columns from a desirability object
}
\keyword{internal}
