% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiples.R
\name{selected_desirability}
\alias{selected_desirability}
\alias{d_max_select}
\title{Desirability functions for selected data columns}
\usage{
d_max_select(.data, ..., low, high, scale = 1, missing = 1)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{...}{One or more unquoted expressions separated by commas (such as
\code{dplyr} selector functions). Variable names can be used as if they were
positions in the data frame, so expressions like \code{x:y} can be used to select
a range of variables.}

\item{low}{Single numeric values that define the active ranges of
desirability.}

\item{high}{Single numeric values that define the active ranges of
desirability.}

\item{scale}{A single numeric value to rescale the
desirability function.}

\item{missing}{A single numeric value on \verb{[0, 1]} (or \code{NA_real_}) that
defines how missing values in \code{x} are mapped to the desirability score.}
}
\value{
A tibble containing the select columns and their desirability scores.
A prefix is added to the column names based on the type of desirability
function (e.g., column \code{x} becomes column \code{d_max_x} when used with
\code{d_max_select()}).
}
\description{
These functions conducts the same desirability computations on multiple
columns of a data set.
}
\examples{
library(dplyr)

set.seed(1)
tibble(x = sort(runif(10)), y = sort(runif(10)), z = 1:10) \%>\%
  d_max_select(x, y, low = 0.5, high = 0.75)
}
