% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-line.R
\name{inline_desirability}
\alias{inline_desirability}
\alias{d_max}
\alias{d_min}
\title{Desirability functions for in-line computations}
\usage{
d_max(x, low, high, scale = 1, missing = 1)

d_min(x, low, high, scale = 1, missing = 1)
}
\arguments{
\item{x}{A vector of data to compute the desirability function}

\item{low, high}{Single numeric values that define the active ranges of
desirability.}

\item{scale}{A single numeric value to rescale the desirability function.}

\item{missing}{A single numeric value on \verb{[0, 1]} (or \code{NA_real_}) that
defines how missing values in \code{x} are mapped to the desirability score.}
}
\value{
A numeric vector on \verb{[0, 1]} where larger values are more
desirable.
}
\description{
Desirability functions map some input to a \verb{[0, 1]} scale where zero is
unacceptable and one is most desirable. The mapping depends on the situation.
\code{d_max()} increases desirability with the input while \code{d_min()} does the
opposite.

Currently, only the desirability functions defined by Derringer and Suich
(1980) are implemented.
}
\examples{
library(dplyr)
library(ggplot2)

set.seed(1)
dat <- tibble(x = sort(runif(10)), y = sort(runif(10)))
d_max(dat$x, 0.1, 0.75)

dat \%>\%
  mutate(d_x = d_max(x, 0.1, 0.75))

set.seed(2)
tibble(z = sort(runif(100))) \%>\%
  mutate(
    no_scale = d_max(z, 0.1, 0.75),
    easier   = d_max(z, 0.1, 0.75, scale = 1/2)
  ) \%>\%
  ggplot(aes(x = z)) +
  geom_point(aes(y = no_scale)) +
  geom_line(aes(y = no_scale), alpha = .5) +
  geom_point(aes(y = easier), col = "blue") +
  geom_line(aes(y = easier), col = "blue", alpha = .5) +
  lims(x = 0:1, y = 0:1) +
  coord_fixed() +
  ylab("Desirability")
}
\references{
Derringer, G. and Suich, R. (1980), Simultaneous Optimization of
Several Response Variables. \emph{Journal of Quality Technology}, 12, 214-219.
}
\seealso{
\code{\link[=d_max_select]{d_max_select()}}, \code{\link[=d_overall]{d_overall()}}
}
